Develop

+++ src/infra/relational/repository/history_project_repository.py
@@ -2,9 +2,17 @@ from src.infra.relational.models.history_project import HistoryProject
 from src.infra.relational.config.interface.i_db_connection_handler import IDBConnectionHandler
 from src.data.interface.i_history_project_repository import IHistoryProjectRepository
 from src.domain.entities.history_project import HistoryProjectEntity
-from src.errors.repository.project_id_not_exists import ProjectIdNotExistsError
-from sqlalchemy.exc import IntegrityError
 from typing import Optional, Dict, List
+
+# Errors
+from src.errors.repository.not_exists_error.project_not_exists import ProjectNotExistsError
+from src.errors.repository.not_exists_error.history_project_not_exists import HistoryProjectNotExists
+from src.errors.repository.error_on_insert.error_on_insert_history_project import ErrorOnInsertHistoryProject
+from src.errors.repository.error_on_find.error_on_find_history_project import ErrorOnFindHistoryProject
+from src.errors.repository.error_on_update.error_on_update_history_project import ErrorOnUpdateHistoryProject
+from src.errors.repository.error_on_delete.error_on_delete_history_project import ErrorOnDeleteHistoryProject
+from src.errors.repository.has_related_children.history_project_has_related_children import HistoryProjectHasRelatedChildren
+from sqlalchemy.exc import IntegrityError
 
 class HistoryProjectRepository(IHistoryProjectRepository):
 
@@ -23,9 +31,11 @@ class HistoryProjectRepository(IHistoryProjectRepository):
                 )
                 db.session.commit()
             except IntegrityError as e:
-                raise ProjectIdNotExistsError(
-                    message=f'Project {project_id} does not exists: {e}'
+                raise ProjectNotExistsError(
+                    message=f'Project {project_id} does not exists: {str(e)}'
                 ) from e
+            except Exception as e:
+                raise ErrorOnInsertHistoryProject(message=f'Error on insert history project {project_id}: {str(e)}') from e
     
     def find(self, history_project_id:int) -> HistoryProjectEntity:
         with self.__db_connection_handler as db:
@@ -33,6 +43,8 @@ class HistoryProjectRepository(IHistoryProjectRepository):
                 history_project = db.session.query(HistoryProject).where(
                     HistoryProject.id == history_project_id
                 ).first()
+                if not history_project:
+                    raise HistoryProjectNotExists(message=f'History project {history_project_id} does not exist')
                 return HistoryProjectEntity(
                     history_project_id = history_project.id,
                     project_id = history_project.project_id,
@@ -40,8 +52,10 @@ class HistoryProjectRepository(IHistoryProjectRepository):
                     description = history_project.description,
                     updated_at = history_project.updated_at
                 )
+            except HistoryProjectNotExists as e:
+                raise e from e
             except Exception as e:
-                raise e
+                raise ErrorOnFindHistoryProject(message=f'Error on find history project {history_project_id}: {str(e)}') from e
     
     def find_all_from_project(self, project_id:int) -> List[HistoryProjectEntity]:
         with self.__db_connection_handler as db:
@@ -49,6 +63,8 @@ class HistoryProjectRepository(IHistoryProjectRepository):
                 full_history = db.session.query(HistoryProject).where(
                     HistoryProject.project_id == project_id
                 ).all()
+                if not any(full_history):
+                    raise HistoryProjectNotExists(message=f'History from project {project_id} does not exist')
                 return [
                     HistoryProjectEntity(
                         history_project_id=register.id,
@@ -59,34 +75,59 @@ class HistoryProjectRepository(IHistoryProjectRepository):
                     ) for register in full_history
                 ]
             except Exception as e:
-                raise e
+                raise ErrorOnFindHistoryProject(message=f'Error on find all history from project {project_id}: {str(e)}') from e
     
     def update(self, history_project_id:int, update_params:Dict) -> None:
         with self.__db_connection_handler as db:
             try:
+                history_project = db.session.query(HistoryProject).where(
+                    HistoryProject.id == history_project_id
+                ).first()
+                if not history_project:
+                    raise HistoryProjectNotExists(f'History project not exists {history_project_id}')
                 db.session.query(HistoryProject).where(
                     HistoryProject.id == history_project_id
                 ).update(update_params)
                 db.session.commit()
+            except HistoryProjectNotExists as e:
+                raise e from e
             except Exception as e:
-                raise e
+                raise ErrorOnUpdateHistoryProject(message=f'Error on update history_project {update_params}: {str(e)}') from e
     
     def delete(self, history_project_id:int) -> None:
         with self.__db_connection_handler as db:
             try:
+                history_project = db.session.query(HistoryProject).where(
+                    HistoryProject.id == history_project_id
+                ).first()
+                if not history_project:
+                    raise HistoryProjectNotExists(f'History project not exists {history_project_id}')
                 db.session.query(HistoryProject).where(
                     HistoryProject.id == history_project_id
                 ).delete()
                 db.session.commit()
+            except HistoryProjectNotExists as e:
+                raise e from e
+            except IntegrityError as e:
+                raise HistoryProjectHasRelatedChildren(message=f'History project {history_project_id} has related children: {str(e)}') from e
             except Exception as e:
-                raise e
+                raise ErrorOnDeleteHistoryProject(message=f'Error on delete history project {history_project_id}: {str(e)}') from e
     
     def delete_all_from_project(self, project_id:int) -> None:
         with self.__db_connection_handler as db:
             try:
+                history_project = db.session.query(HistoryProject).where(
+                    HistoryProject.project_id == project_id
+                ).all()
+                if not any(history_project):
+                    raise HistoryProjectNotExists(f'History project from project {project_id} not exists')
                 db.session.query(HistoryProject).where(
                     HistoryProject.project_id == project_id
                 ).delete()
                 db.session.commit()
-            except Exception as e:
+            except HistoryProjectNotExists as e:
                 raise e from e
+            except IntegrityError as e:
+                raise HistoryProjectHasRelatedChildren(message=f'History project from project {project_id} has related children: {str(e)}') from e
+            except Exception as e:
+                raise ErrorOnDeleteHistoryProject(message=f'Error on delete all history from project {project_id}: {str(e)}') from e
