+++ src/infra/relational/repository/project_type_repository.py
@@ -1,6 +1,5 @@
 #pylint:disable=W0611
 from typing import List
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy import and_
 
 from src.infra.relational.models.project_type import ProjectType
@@ -8,13 +7,18 @@ from src.infra.relational.config.interface.i_db_connection_handler import IDBCon
 from src.data.interface.i_project_type_repository import IProjectTypeRepository
 from src.domain.entities.project_type import ProjectTypeEntity
 
-from src.errors.repository.project_type_already_exists import ProjectTypeAlreadyExists
-from src.errors.repository.projects_from_type_does_not_exists import ProjectsFromTypeDoesNotExists
-from src.errors.repository.error_on_update_type_from_project import ErrorOnUpdateTypeFromProject
-from src.errors.repository.error_on_delete_project_type import ErrorOnDeleteProjectType
-
 from src.infra.relational.models.types import Types
 from src.infra.relational.models.project import Project
+
+# Errors
+from src.errors.repository.already_exists_error.project_type_already_exists import ProjectTypeAlreadyExists
+from src.errors.repository.not_exists_error.project_type_not_exists import ProjectTypeNotExists
+from src.errors.repository.error_on_delete.error_on_delete_project_type import ErrorOnDeleteProjectType
+from src.errors.repository.error_on_find.error_on_find_project_type import ErrorOnFindProjectType
+from src.errors.repository.error_on_insert.error_on_insert_project_type import ErrorOnInsertProjectType
+from src.errors.repository.error_on_update.error_on_update_project_type import ErrorOnUpdateProjectType
+from src.errors.repository.has_related_children.project_type_has_related_children import ProjectTypeHasRelatedChildren
+from sqlalchemy.exc import IntegrityError
 
 class ProjectTypeRepository(IProjectTypeRepository):
     def __init__(self, db_connection_handler: IDBConnectionHandler) -> None:
@@ -30,7 +34,9 @@ class ProjectTypeRepository(IProjectTypeRepository):
                 message=f'ProjectType with (project_id: {project_id}, type_id: {type_id}) already exists: {e}'
             ) from e
         except Exception as e:
-            raise e
+            raise ErrorOnInsertProjectType(
+                message=f'Error on insert project type project_id={project_id}, type_id={type_id}: {str(e)}'
+            ) from e
 
     def find(self, project_id: int, type_id: int) -> ProjectTypeEntity:
         try:
@@ -42,7 +48,7 @@ class ProjectTypeRepository(IProjectTypeRepository):
                     )
                 ).first()
                 if not relation:
-                    raise ProjectsFromTypeDoesNotExists(
+                    raise ProjectTypeNotExists(
                         message=f'No ProjectType with (project_id: {project_id}, type_id: {type_id}) found'
                     )
                 return ProjectTypeEntity(
@@ -50,8 +56,13 @@ class ProjectTypeRepository(IProjectTypeRepository):
                     type_id=relation.type_id,
                     created_at=relation.created_at
                 )
+        except ProjectTypeNotExists as e:
+            raise e from e
         except Exception as e:
-            raise e
+            raise ErrorOnFindProjectType(
+                message=f'Error on find project type project_id={project_id}, type_id={type_id}: {str(e)}'
+            ) from e
+
 
     def find_all_from_type(self, type_id: int) -> List[ProjectTypeEntity]:
         try:
@@ -60,7 +71,7 @@ class ProjectTypeRepository(IProjectTypeRepository):
                     ProjectType.type_id == type_id
                 ).all()
                 if not relations:
-                    raise ProjectsFromTypeDoesNotExists(
+                    raise ProjectTypeNotExists(
                         message=f'No projects found for type_id: {type_id}'
                     )
                 return [
@@ -70,12 +81,26 @@ class ProjectTypeRepository(IProjectTypeRepository):
                         created_at=relation.created_at
                     ) for relation in relations
                 ]
+        except ProjectTypeNotExists as e:
+            raise e from e
         except Exception as e:
-            raise e
+            raise ErrorOnFindProjectType(
+                message=f'Error on find project_type with type_id={type_id}: {str(e)}'
+            ) from e
 
     def update_type(self, project_id: int, type_id: int, new_type_id: int) -> None:
         try:
             with self.__db_connection_handler as db:
+                project_type = db.session.query(ProjectType).where(
+                    and_(
+                        ProjectType.project_id == project_id,
+                        ProjectType.type_id == type_id
+                    )
+                ).first()
+                if not project_type:
+                    raise ProjectTypeNotExists(
+                        message=f'Project type project_id={project_id}, type_id={type_id} not exists'
+                    )
                 db.session.query(ProjectType).where(
                     and_(
                         ProjectType.project_id == project_id,
@@ -83,14 +108,30 @@ class ProjectTypeRepository(IProjectTypeRepository):
                     )
                 ).update({'type_id': new_type_id})
                 db.session.commit()
+        except ProjectTypeNotExists as e:
+            raise e from e
+        except IntegrityError as e:
+            raise ProjectTypeAlreadyExists(
+                message=f'Project type project_id={project_id}, type_id={new_type_id} already exists'
+            ) from e
         except Exception as e:
-            raise ErrorOnUpdateTypeFromProject(
+            raise ErrorOnUpdateProjectType(
                 message=f'Error updating ProjectType (project_id: {project_id}, type_id: {type_id}) to {new_type_id}'
             ) from e
 
     def delete(self, project_id: int, type_id: int) -> None:
         try:
             with self.__db_connection_handler as db:
+                project_type = db.session.query(ProjectType).where(
+                    and_(
+                        ProjectType.project_id == project_id,
+                        ProjectType.type_id == type_id
+                    )
+                ).first()
+                if not project_type:
+                    raise ProjectTypeNotExists(
+                        message=f'Project type project_id={project_id}, type_id={type_id} not exists'
+                    )
                 db.session.query(ProjectType).where(
                     and_(
                         ProjectType.project_id == project_id,
@@ -98,6 +139,12 @@ class ProjectTypeRepository(IProjectTypeRepository):
                     )
                 ).delete()
                 db.session.commit()
+        except ProjectTypeNotExists as e:
+            raise e from e
+        except IntegrityError as e:
+            raise ProjectTypeHasRelatedChildren(
+                message=f'Project type from project project_id={project_id}, type_id={type_id} has related children'
+            ) from e
         except Exception as e:
             raise ErrorOnDeleteProjectType(
                 message=f'Error deleting ProjectType (project_id: {project_id}, type_id: {type_id}): {e}'
@@ -106,10 +153,23 @@ class ProjectTypeRepository(IProjectTypeRepository):
     def delete_all_from_project(self, project_id: int) -> None:
         try:
             with self.__db_connection_handler as db:
+                project_type = db.session.query(ProjectType).where(
+                    ProjectType.project_id == project_id,
+                ).all()
+                if not any(project_type):
+                    raise ProjectTypeNotExists(
+                        message=f'Project Type from project project_id={project_id} not exists'
+                    )
                 db.session.query(ProjectType).where(
                     ProjectType.project_id == project_id,
                 ).delete()
                 db.session.commit()
+        except ProjectTypeNotExists as e:
+            raise e from e
+        except IntegrityError as e:
+            raise ProjectTypeHasRelatedChildren(
+                message=f'Project type from project_id={project_id} has related children: {str(e)}'
+            ) from e
         except Exception as e:
             raise ErrorOnDeleteProjectType(
                 message=f'Error deleting ProjectType (project_id: {project_id}): {e}'
@@ -121,6 +181,10 @@ class ProjectTypeRepository(IProjectTypeRepository):
                 relations = db.session.query(ProjectType).where(
                     ProjectType.project_id == project_id,
                 ).all()
+                if not any(relations):
+                    raise ProjectTypeNotExists(
+                        message=f'Project type from project project_id={project_id} does not exists'
+                    )
                 return [
                     ProjectTypeEntity(
                         project_id=relation.project_id,
@@ -128,5 +192,9 @@ class ProjectTypeRepository(IProjectTypeRepository):
                         created_at=relation.created_at
                     ) for relation in relations
                 ]
-        except Exception as e:
+        except ProjectTypeNotExists as e:
             raise e from e
+        except Exception as e:
+            raise ErrorOnFindProjectType(
+                message=f'Error on find all from project project_id={project_id}: {str(e)}'
+            ) from e
